requirements.txt
Flask==1.0.2
SQLAlchemy==1.3.3
alembic==1.0.10
beautifulsoup4==4.8.1
factory_boy==2.11.1
jaeger_client==3.13.0
jsonrpcserver==4.0.2
numpy==1.17.3
opentracing_instrumentation==2.4.3
paramiko==2.7.2
pika==1.0.1
psycopg2-binary==2.8.2
pylzma==0.5.0
pytest==5.2.2
python-dotenv==0.10.3
python-logstash-async==1.4.1
python-logstash==0.4.6
redis==3.0.1
requests==2.22.0
transliterate==1.10.2
